+++ frontend/src/components/ToolsSection/FeatureCards/components/index.ts
@@ -1,0 +1,1 @@
+export { default as FeatureCard } from './FeatureCard';
